# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {
            "name": "Apache 2.0"
        }
    },
    "paths": {
...
        "/api/v1/base/security-levels": {
            "get": {
                "description": "查询安全级别列表",
                "tags": [
                    "SecurityLevel"
                ],
                "summary": "查询安全级别列表",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "是否仅查询count",
                        "name": "-",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "当前页",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否查找所有数据",
                        "name": "find_all",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "安全级别ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "安全级别ID集合(逗号分隔)",
                        "name": "ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "安全级别名称(模糊查询)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页大小",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "模糊搜索(搜索 安全级别名称/备注)",
                        "name": "queryValue",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "角色ID集合(逗号分隔)",
                        "name": "role_ids",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否显示角色信息",
                        "name": "show_details",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态(1:启用 2:禁用)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.ListResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            },
            "post": {
                "description": "创建安全级别(如果绑定的角色没有ID，则创建角色)",
                "tags": [
                    "SecurityLevel"
                ],
                "summary": "创建安全级别(如果绑定的角色没有ID，则创建角色)",
                "parameters": [
                    {
                        "description": "创建安全级别参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.SecurityLevel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.IDResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
        "/api/v1/base/security-levels/:id": {
            "get": {
                "description": "获取安全级别和绑定的角色信息",
                "tags": [
                    "SecurityLevel"
                ],
                "summary": "获取安全级别和绑定的角色信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "安全级别ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.SecurityLevel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            },
            "put": {
                "description": "更新安全级别基础信息和安全级别与角色的关联关系",
                "tags": [
                    "SecurityLevel"
                ],
                "summary": "更新安全级别基础信息和安全级别与角色的关联关系",
                "parameters": [
                    {
                        "description": "更新安全级别信息参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.SecurityLevel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            },
            "delete": {
                "description": "删除安全级别",
                "tags": [
                    "SecurityLevel"
                ],
                "summary": "删除安全级别",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "安全级别ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
        "/api/v1/base/system-config": {
            "get": {
                "description": "查询系统配置信息列表",
                "tags": [
                    "SystemConfig"
                ],
                "summary": "查询系统配置信息列表",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "是否仅查询count",
                        "name": "-",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "当前页",
                        "name": "current",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页大小",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.ListResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
        "/api/v1/base/system-config/:id": {
            "put": {
                "description": "更新系统基础配置信息",
                "tags": [
                    "SystemConfig"
                ],
                "summary": "更新系统基础配置信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "系统基础配置ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新系统基础配置信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.SystemConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ginx.OperateResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.IDResult": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
        "common.PaginationResult": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "ginx.ListResult": {
            "type": "object",
            "properties": {
                "list": {},
                "pagination": {
                    "$ref": "#/definitions/common.PaginationResult"
                }
            }
        },
        "ginx.OperateResult": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
...
        "schema.Role": {
            "type": "object",
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "auditor_types": {
                    "description": "审核类型(逗号分隔)",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "integer"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "角色名称",
                    "type": "string"
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "status": {
                    "description": "状态(1:启用 2:禁用)",
                    "type": "integer"
                },
                "type": {
                    "description": "角色类型(1.用户角色 2.组织角色 3.职位角色 4.用户组角色)",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
        "schema.SecurityLevel": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "integer"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "integer"
                },
                "memo": {
                    "description": "备注",
                    "type": "string"
                },
                "name": {
                    "description": "安全级别名称",
                    "type": "string"
                },
                "roles": {
                    "description": "安全级别绑定的角色",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Role"
                    }
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                },
                "status": {
                    "description": "状态(1:启用 2:禁用)",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
        "schema.SystemConfig": {
            "type": "object",
            "required": [
                "web_server_host",
                "web_server_port"
            ],
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "https_crt_file": {
                    "description": "Https Crt证书",
                    "type": "string"
                },
                "https_key_file": {
                    "description": "Https私钥",
                    "type": "string"
                },
                "https_mode": {
                    "description": "是否启用Https",
                    "type": "boolean"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "integer"
                },
                "menu_version": {
                    "description": "前端菜单版本号(只能通过定时任务，由菜单json中的版本号字段修改)",
                    "type": "number"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "web_server_host": {
                    "description": "Web服务IP地址",
                    "type": "string"
                },
                "web_server_port": {
                    "description": "Web服务端口",
                    "type": "integer"
                }
            }
        },
...
    }
}